move method minimum_outdegree_orientation
